rename globals
