+++ base_station/static/globalvars.py
